requirements.txt
streamlit==1.34.0
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
openai==0.28.0